re-use the code found in the json importer babel transformer
