+++ build-system/babel-plugins/babel-plugin-transform-json-import/index.js
@@ -1,5 +1,5 @@
 const {dirname, join, relative, resolve} = require('path');
-const {readFileSync} = require('fs');
+const {readJson} = require('../../json-locales');
 
 /**
  * Transforms JSON imports into a `JSON.parse` call:
@@ -19,32 +19,6 @@ const {readFileSync} = require('fs');
  */
 module.exports = function (babel) {
   const {template, types: t} = babel;
-
-  /**
-   * JSON reviver that converts {"string": "foo", ...} to just "foo".
-   * This minifies the format of locale files.
-   * @param {string} _
-   * @param {*} value
-   * @return {*}
-   */
-  function minifyLocalesJsonReviver(_, value) {
-    // Always default to original `value` since this reviver is called for any
-    // property pair, including the higher-level containing object.
-    return value?.string || value;
-  }
-
-  /**
-   * @param {string} filename
-   * @return {*}
-   */
-  function readJson(filename) {
-    // Treat files under /_locales/ specially in order to minify their format.
-    const reviver = filename.includes('/_locales/')
-      ? minifyLocalesJsonReviver
-      : undefined;
-    return JSON.parse(readFileSync(filename, 'utf8'), reviver);
-  }
-
   return {
     manipulateOptions(_opts, parserOpts) {
       parserOpts.plugins.push('importAssertions');
